Metrics.jl

Metrics.Classwise_Stats

Metrics.Perplexity

Metrics.TFPN

Metrics._f_p_r_lcs

Metrics._get_ngrams

Metrics._get_word_ngrams

Metrics._lcs

Metrics._len_lcs

Metrics._recon_lcs

Metrics._split_into_words

Metrics._union_lcs

Metrics.adjusted_r2_score

Metrics.avg_precision

Metrics.binary_accuracy

Metrics.bleu_score

Metrics.categorical_accuracy

Metrics.cohen_kappa

Metrics.confusion_matrix

Metrics.f_beta_score

Metrics.false_alarm_rate

Metrics.get_ngrams

Metrics.global_stats

Metrics.mae

Metrics.male

Metrics.mse

Metrics.msle

Metrics.precision

Metrics.r2_score

Metrics.ranking_stats_k

Metrics.recall

Metrics.report_stats

Metrics.rouge

Metrics.rouge_l_sentence_level

Metrics.rouge_l_summary_level

Metrics.rouge_n

Metrics.sparse_categorical

Metrics.specificity

Metrics.statsfromTFPN

Metrics.top_k_categorical

Metrics.top_k_sparse_categorical

Metrics.Classwise_Stats — Method

classwise_stats(y_pred, y_true)

Computes statistics for each of the class for multiclass classification based on provided y_pred and y_true.

Return the result stats as a dictionary.

Metrics.Perplexity — Method

perplexity(y_pred, y_true)

Returns the exponentiation of crossentropy based on y_pred and y_true.

Metrics.TFPN — Method

TFPN(y_pred, y_true)

Returns Confusion Matrix and True Positive, True Negative, False Positive and False Negative for each class based on y_pred and y_true. Expects y_true, to be onehot_enocded already.

Metrics._f_p_r_lcs — Method

_f_p_r_lcs(llcs, m, n)

Computes the LCS-based F-measure score

Arguments:

llcs: Length of LCS

m: number of words in reference summary

n: number of words in candidate summary

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

Metrics._get_ngrams — Method

_get_ngrams(n, text)

Calcualtes n-grams. Returns a set of n-grams.

Arguments:

n: provide which n-grams to calculate

text: An array of tokens

Metrics._get_word_ngrams — Method

_get_word_ngrams(n, sentences)

Calculates word n-grams for multiple sentences.

Metrics._lcs — Method

_lcs(x, y)

Utility function to compute the length of the longest common subsequence (lcs) between two strings. The implementation below uses a DP programming algorithm and runs in O(nm) time where n = len(x) and m = len(y).

Metrics._len_lcs — Method

_len_lcs(x, y)

Returns the length of the Longest Common Subsequence between sequences x and y.

Metrics._recon_lcs — Method

_recons_lcs(x, y)

Returns the Longest Subsequence between x and y.

Metrics._split_into_words — Method

_split_into_words(sentences)

Splits multiple sentences into words and flattens the result

Metrics._union_lcs — Method

_union_lcs(evaluated_sentences, reference_sentence)

Returns LCSu(ri, C) which is the LCS score of the union longest common subsequence between reference sentence ri and candidate summary C.

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentence: one of the sentences in the reference summaries

For example, if ri= w1 w2 w3 w4 w5, and C contains two sentences: c1 = w1 w2 w6 w7 w8 and c2 = w1 w3 w8 w9 w5, then the longest common subsequence of ri and c1 is “w1 w2” and the longest common subsequence of ri and c2 is “w1 w3 w5”. The union longest common subsequence of ri, c1, and c2 is “w1 w2 w3 w5” and LCSu(ri, C) = 4/5.

Metrics.adjusted_r2_score — Method

adjusted_r2_score(y_pred, y_true, n)

Modified version of r2_score that has been adjusted for the number of predictors in the model. Here the argument n is for the number of predictors(or independent variables in X).

See also: r2_score

Metrics.avg_precision — Function

avg_precision(y_rec, y_rel, k = 10)

Evaluates how much of the relevant documents are concentrated in the highest ranked predictions.

Calculated as ∑(Recall@i - Recall@i-1)* Precision@i for i = (1, 2, 3....k)

Here, y_rec are predicted probabilities for recommendation and y_rel defines as 1 if particular result is relevant, else 0. The shape of y_rec is expected to be (1, N_elements)

Metrics.binary_accuracy — Method

binary_accuracy(y_pred, y_true; threshold=0.5)

Calculates Averaged Binary Accuracy based on y_pred and y_true. Argument threshold is used to specify the minimum predicted probability y_pred required to be labelled as 1. Default value set as 0.5.

Metrics.bleu_score — Method

bleu_score(reference_corpus, translation_corpus; max_order=4, smooth=false)

Computes BLEU score of translated segments against one or more references. Returns the BLEU score, n-gram precisions, brevity penalty, geometric mean of n-gram precisions, translationlength and referencelength

Arguments

reference_corpus: list of lists of references for each translation. Each reference should be tokenized into a list of tokens.

translation_corpus: list of translations to score. Each translation should be tokenized into a list of tokens.

max_order: maximum n-gram order to use when computing BLEU score.

smooth=false: whether or not to apply. Lin et al. 2004 smoothing.

Metrics.categorical_accuracy — Method

categorical_accuracy(y_pred, y_true)

Calculates Averaged Categorical Accuracy based on y_pred and y_true.

Metrics.cohen_kappa — Method

cohen_kappa(y_pred, y_true)

Measures the agreement between two raters (predicted and ground truth, here) who each classify N items into C mutually exclusive categories, using the observed data to calculate the probabilities of each observer randomly seeing each category. If the raters are in complete agreement then κ = 1. If there is no agreement among the raters other than what would be expected by chance, κ = 0.

Ref: Cohen's Kappa

Metrics.confusion_matrix — Method

confusion_matrix(y_pred, y_true)

Function to create a confusionmatrix for classification problems based on provided `ypredandytrue. Expectsytrue`, to be onehot_enocded already.

Metrics.f_beta_score — Method

f_beta_score(y_pred, y_true; β=1, avg_type="macro", sample_weights=nothing)

Compute fbeta score. The F_beta score is the weighted harmonic mean of precision and recall, reaching its optimal value at 1 and its worst value at 0.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed.

β=1: the weight of precision in the combined score. If β<1, more weight given to precision, while β>1 favors recall.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Metrics.false_alarm_rate — Method

false_alarm_rate(y_pred, y_true; avg_type="macro", sample_weights=nothing)

Computes the falsealarmraye of the predictions with respect to the labels as 1 - specificity(y_pred, y_true, avg_type, sample_weights)

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

See also: specificity

Metrics.get_ngrams — Method

get_ngrams(segment, max_order)

Extracts all n-grams upto a given maximum order from an input segment. Returns the counter containing all n-grams upto max_order in segment with a count of how many times each n-gram occurred.

Arguments

segment: text segment from which n-grams will be extracted.

max_order: maximum length in tokens of the n-grams returned by this methods.

Metrics.global_stats — Method

global_stats(y_pred, y_true; avg_type="macro")

Computes the overall statistics based on provided y_pred and y_true. avg_type allows to specify the type of average to be used while evaluating the stats. Currently, it can take values as "macro" or "micro".

Return the result stats as a dictionary.

Metrics.mae — Method

mae(y_pred, y_true)

Mean Absolute Error. Calculated as sum(|y_true .- y_pred|) / length(y_true) based on provided y_pred and y_true.

Metrics.male — Method

male(y_pred, y_true)

Mean Absolute Logarithmic Error. Calculated as sum(|log.(y_true) .- log.(y_pred)|) / length(y_true) based on provided y_pred and y_true.

Metrics.mse — Method

mse(y_pred, y_true)

Mean Squared Error. Calculated as sum((y_true .- y_pred).^2) / length(y_true) based on provided y_pred and y_true.

Metrics.msle — Method

msle(y_pred, y_true)

Mean Absolute Logarithmic Error. Calculated as sum((log.(y_true) .- log.(y_pred)).^2) / length(y_true) based on provided y_pred and y_true.

Metrics.precision — Method

precision(y_pred, y_true; avg_type="macro", sample_weights=nothing)

Computes the precision of the predictions with respect to the labels.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Metrics.r2_score — Method

r2_score(y_pred, y_true)

Calculates the r2 (Coefficient of Determination) score for the provided y_pred and y_true. Best possible score is 1.0 and it can be negative (because the model can be arbitrarily worse). A constant model that always predicts the expected value of y, disregarding the input features, would get a r2_score of 0.0.

Metrics.ranking_stats_k — Function

ranking_stats_k(y_rec, y_rel, k = 10)

Evaluates the relevancy of top k recommendations using precison@k, recall@k and f1_score@k. Returns result as a Dict.

Here, y_rec are predicted probabilities for recommendation and y_rel defines as 1 if particular result is relevant, else 0. The shape of y_rec is expected to be (1, N_elements).<br>

precison_k is evaluated as Recommended_items_that_are_relevant / Total_Recommended_items recall_l is evaluated as Recommended_items_that_are_relevant / Total_Relevant_items f1_k is evaluated as 2 * Recommended_items_that_are_relevant / (Total_Recommended_items + Total_Relevant_items)

Metrics.recall — Method

recall(y_pred, y_true; avg_type="macro", sample_weights=nothing)

Computes the recall of the predictions with respect to the labels.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Aliases: sensitivity and detection_rate

Metrics.report_stats — Method

report_stats(y_pred, y_true; classwise_stats=true, avg_type="macro", sample_weights=nothing)

A utility function that prints the statistics summary of the model based on provided y_pred and y_true.

Arguments:

y_pred: predicted values

y_true: ground truth values on the basis of which predicted values are to be assessed.

classwise_stats=true: if set true, prints classwise stats along with global stats.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Metrics.rouge — Method

rouge(hypotheses, references)

Calculates average rouge scores for a list of hypotheses and references.

Metrics.rouge_l_sentence_level — Method

rouge_l_sentence_level(evaluated_sentences, reference_sentences)

Computes ROUGE-L (sentence level) of two text collections of sentences.

Calculated according to: Rlcs = LCS(X,Y)/m, Plcs = LCS(X,Y)/n, Flcs = ((1 + beta^2)*Rlcs*Plcs) / (Rlcs + (beta^2) * P_lcs)

where: X = reference summary Y = Candidate summary m = length of reference summary n = length of candidate summary

Argumnets:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentences: the sentences from the referene set

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

Metrics.rouge_l_summary_level — Method

rouge_l_summary_level(evaluated_sentences, reference_sentences)

Computes ROUGE-L (summary level) of two text collections of sentences.

Calculated according to: Rlcs = SUM(1, u)[LCS<union>(ri,C)]/m Plcs = SUM(1, u)[LCS<union>(ri,C)]/n Flcs = ((1 + beta^2)*Rlcs*Plcs) / (Rlcs + (beta^2) * P_lcs)

where: SUM(i,u) = SUM from i through u u = number of sentences in reference summary C = Candidate summary made up of v sentences m = number of words in reference summary n = number of words in candidate summary

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentence: the sentences in the reference summaries

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

Metrics.rouge_n — Method

rouge_n(evaluated_sentences, reference_sentences; n=2)

Computes ROUGE-N of two text collections of sentences. Returns f1, precision, recall for ROUGE-N.

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentences: the sentences from the referene set

n: size of ngram. Defaults to 2.

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/ papers/rouge-working-note-v1.3.1.pdf)

Metrics.sparse_categorical — Method

sparse_categorical(y_pred, y_true)

Calculated Sparse Categorical Accuracy based on y_pred and y_true. It evaluates the maximal true value is equal to the index of the maximal predicted value. Here, y_true is expected to provide only an integer as label for each data element (ie. not one hot encoded).

Metrics.specificity — Method

specificity(y_pred, y_true; avg_type="macro", sample_weights=nothing)

Computes the specificity of the predictions with respect to the labels.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Metrics.statsfromTFPN — Method

statsfromTFPN(TP, TN, FP, FN)

Computes statistics in case of binary classification or one-vs-all statsitics in case of multiclass classification.

Arguments:

TP: true positive values

TN: true negative values

FP: false positive values

FN: false negative values

Return the result stats as a dictionary.

Metrics.top_k_categorical — Method

top_k_categorical(y_pred, y_true; k=3)

Evaluates if the index of true value is equal to any of the indices of top k predicted values. Default value of k set to 3.

Metrics.top_k_sparse_categorical — Method

top_k_sparse_categorical(y_pred, y_true; k=3)

Evaluates if the true value is equal to any of the indices of top k predicted values. Default value of k set to 3. Similar to sparse_categorical, expects the y_true to provide only an integer as label for each data element (ie. not one hot encoded).