Computer Vision Metrics

These Metrics are meant to evaluate the Computer Vision Algorithms / Models. Their main purpose include analysing the quality of generated images, extent of accuracy in case of Object Detection and so on.

Functions

Metrics.PSNR — Function

PSNR(img1, img2)

Computes peak-signal-to-noise ratio, in decibels, between two images img1 and img2. The higher the PSNR, the better the quality of the compressed, or reconstructed image.

Metrics.IoU — Function

IoU(bb1, bb2)

Calculate the Intersection over Union (IoU) of two axis-aligned bounding boxes bb1 and bb2.

Here, bb1 and bb2 are provided as Dict with keys = {"x1", "x2", "y1", "y2"}, where x1, y1 are coordinates of top-left corner, and x2 and y2 are coordinates of bottom-right corner.